+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## Change Log
 
+### Ver 0.0.2
+
+  * #2 [Enable to apply custom style on preview items](https://github.com/georgeOsdDev/react-dropfile-field/issues/2)
+  * #3 [Enable to limit files length](https://github.com/georgeOsdDev/react-dropfile-field/issues/3 )
+
 ### Ver 0.0.1 Initial release